requirements.txt
python-dotenv==1.2.1
grpcio>=1.74.0
grpcio-health-checking>=1.74.0
protobuf>=5.27.0
